refactor(agent): restructure into core/ and runtime/ directories
