+++ packages/agent/src/core/budget.ts
