+++ packages/agent/src/core/index.ts
@@ -1,0 +1,10 @@
+export { ChatAgent, type ChatAgentInstance } from "./chat-agent";
+export type {
+  ChatAgentConfig,
+  ChatAgentInput,
+  AgentResult,
+  AgentStep,
+  AgentBudget,
+  TokenUsage,
+  Sink,
+} from "./types";
